requirements.txt
fastapi
redis
uvicorn
pydantic
pyyaml
requests
python-dotenv
groq
prometheus_client
scikit-learn
tqdm
joblib
numpy